refactor progress observable from list creation operator
